pyproject.toml
[project]
name = "tts-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.10",
    "confluent-kafka>=2.11.0",
    "kokoro>=0.9.4",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "soundfile>=0.13.1",
    "openai-whisper>=20250625",
]
